+++ CHANGELOG.md
@@ -24,6 +24,8 @@ All notable changes to this project will be documented in this file.
 - JpegExporter for OxyPlot.ImageSharp
 - Multi-line aligned text rendering example
 - WindowsForms ExampleBrowser can display transposed versions of examples (#1535)
+- Example for Issue #1545 showing the use of different line endings
+- Support for unix line endings in OxyPlot.ImageSharp, OxyPlot.Svg, and OxyPlot.Pdf (#1545)
 
 ### Changed
 - Legends model (#644)